+++ src/settings/Aggregators/AggregatorConfig/AggregatorConfigForm.js
@@ -1,5 +1,7 @@
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
+import { FieldArray } from 'react-final-form-arrays';
+import { Field } from 'react-final-form';
 import {
   Col,
   InfoPopover,
@@ -8,58 +10,83 @@ import {
   TextField,
 } from '@folio/stripes/components';
 
-function AggregatorConfigForm({ fields, onAddField, onChange, onRemoveField }) {
+import { required } from '../../../util/validate';
+
+function AggregatorConfigForm({ stripes }) {
   const intl = useIntl();
+  const disabled = !stripes.hasPerm('ui-erm-usage.generalSettings.manage');
 
   return (
-    <>
-      <InfoPopover
-        content={`${intl.formatMessage({
-          id: 'ui-erm-usage.aggregator.config.popover',
-        })} apiKey, requestorId, customerId, reportRelease`}
-      />
-      <RepeatableField
-        id="add-agg-config-param"
-        addLabel={intl.formatMessage({
-          id: 'ui-erm-usage.aggregator.config.addParam',
-        })}
-        fields={fields}
-        onAdd={onAddField}
-        onRemove={onRemoveField}
-        renderField={(field, index) => (
+    <FieldArray name="aggregatorConfig">
+      {({ fields }) => (
+        <>
+          <InfoPopover
+            content={`${intl.formatMessage({
+              id: 'ui-erm-usage.aggregator.config.popover',
+            })} apiKey, requestorId, customerId, reportRelease`}
+          />
           <Row>
-            <Col xs={6} sm={4}>
-              <TextField
-                autoFocus
-                label="Key"
-                name={`agg[${index}].key`}
-                id={`aggregator-conf-input-key-${index}`}
-                onChange={(e) => onChange('key', index, e)}
-                value={field.key}
-                disabled={field.isInitial}
-              />
-            </Col>
-            <Col xs={6} sm={4}>
-              <TextField
-                label="Value"
-                name={`agg[${index}].value`}
-                id={`aggregator-conf-input-value-${index}`}
-                onChange={(e) => onChange('value', index, e)}
-                value={field.value}
+            <Col xs={8}>
+              <RepeatableField
+                id="add-agg-config-param"
+                addLabel={intl.formatMessage({
+                  id: 'ui-erm-usage.aggregator.config.addParam',
+                })}
+                fields={fields.value || []}
+                onAdd={() => fields.push({ key: '', value: '', isInitial: false })}
+                onRemove={(index) => fields.remove(index)}
+                renderField={(field, index) => (
+                  <Row>
+                    <Col xs={6}>
+                      <Field
+                        name={`aggregatorConfig[${index}].key`}
+                        validate={required}
+                      >
+                        {({ input, meta }) => (
+                          <TextField
+                            {...input}
+                            autoFocus
+                            label="Key"
+                            id={`aggregator-conf-input-key-${index}`}
+                            disabled={field.isInitial || disabled}
+                            error={meta.touched ? meta.error : undefined}
+                            required
+                          />
+                        )}
+                      </Field>
+                    </Col>
+                    <Col xs={6}>
+                      <Field
+                        name={`aggregatorConfig[${index}].value`}
+                        validate={required}
+                      >
+                        {({ input, meta }) => (
+                          <TextField
+                            {...input}
+                            label="Value"
+                            id={`aggregator-conf-input-value-${index}`}
+                            disabled={disabled}
+                            error={meta.touched ? meta.error : undefined}
+                            required
+                          />
+                        )}
+                      </Field>
+                    </Col>
+                  </Row>
+                )}
               />
             </Col>
           </Row>
-        )}
-      />
-    </>
+        </>
+      )}
+    </FieldArray>
   );
 }
 
 AggregatorConfigForm.propTypes = {
-  fields: PropTypes.arrayOf(PropTypes.shape()).isRequired,
-  onAddField: PropTypes.func.isRequired,
-  onChange: PropTypes.func.isRequired,
-  onRemoveField: PropTypes.func.isRequired,
+  stripes: PropTypes.shape({
+    hasPerm: PropTypes.func.isRequired,
+  }).isRequired,
 };
 
 export default AggregatorConfigForm;