+++ src/settings/Aggregators/AggregatorDetails.js
@@ -17,7 +17,6 @@ import {
 import { MOD_SETTINGS } from '../../util/constants';
 import DownloadCredentialsButton from './DownloadCredentialsButton';
 import { AggregatorConfigView } from './AggregatorConfig';
-
 import aggregatorAccountConfigTypes from '../../util/data/aggregatorAccountConfigTypes';
 
 const AggregatorDetails = ({
@@ -44,9 +43,9 @@ const AggregatorDetails = ({
   };
 
   const renderContact = (aggregator) => {
-    if (aggregator.accountConfig && aggregator.accountConfig.displayContact) {
-      return aggregator.accountConfig.displayContact.map((item, i) => (
-        <p key={i}>{item}</p>
+    if (aggregator?.accountConfig?.displayContact) {
+      return aggregator.accountConfig.displayContact.map((item) => (
+        <p key={item}>{item}</p>
       ));
     } else {
       return null;
@@ -71,8 +70,6 @@ const AggregatorDetails = ({
 
   const settings = resources?.settings?.records || [];
   const hideValues = settings[0]?.value === true;
-
-  const config = aggregator.aggregatorConfig;
 
   const displayWhenOpenAccountConfAcc = (
     <DownloadCredentialsButton
@@ -125,7 +122,7 @@ const AggregatorDetails = ({
           <Row>
             <Col xs={8}>
               <AggregatorConfigView
-                aggregatorConfig={config}
+                aggregatorConfig={aggregator.aggregatorConfig}
                 hideValues={hideValues}
               />
             </Col>